// frontend/src/api/generated/content/content.ts
export type listContentPostsApiV1ContentGetResponse200 = {
  data: PageContentPostRead;
  status: 200;
};

export type listContentPostsApiV1ContentGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type listContentPostsApiV1ContentGetResponseSuccess =
  listContentPostsApiV1ContentGetResponse200 & {
    headers: Headers;
  };
export type listContentPostsApiV1ContentGetResponseError =
  listContentPostsApiV1ContentGetResponse422 & {
    headers: Headers;
  };

export type listContentPostsApiV1ContentGetResponse =
  | listContentPostsApiV1ContentGetResponseSuccess
  | listContentPostsApiV1ContentGetResponseError;